MFA: Change from permission to a new concept `requirement`
A requirement is a restriction on a resource, an anti permission if you
will. Currently two types exist, `mfa-enforced` and `none`.

Requirements are required to be set on a resource, so that an
implementor can check requirements without having to check if the
attribute is set in the request beforehand.